isolating behaviour, global behaviour required for modularisation. All behaviour unique to index.html needs to remain or be moved to mainModular.js. DONT LET ERRORS DEFINE behaviour locating...
