+++ index.html
@@ -15,8 +15,9 @@
   <script src="https://code.jquery.com/jquery-1.11.0.min.js" defer></script>
   
   <!-- <script src="js/mainNoModul.js" defer></script> -->
-  <script type="module" src="js/navbar.js"></script>
+  <script type="module" defer src="js/navbar.js"></script>
   <script type="module" defer src="js/mainModular.js" ></script>
+  
   <script nomodule defer src="js/mainNonModular.js" ></script>
