+++ js/mainModular.js
@@ -1,16 +1,23 @@
 // mainModular.js
 // import { showWip, hideWip } from "./messagesAndForms.js";
-import { getDomElements, debounce } from "./domUtils.js";
+import { getDomElements, debounce, getEndTopY, updateEndTopY } from "./domUtils.js";
 
 import { prioritizedImages, thumbnailImages, GSAP_DEFAULTS } from "./config.js";
 
 import {preloadImages, preloadThumbnailImages,lazyLoadImages} from "./preload.js";
-import { collectThumbs, SCROLL_DURATION, showStatementContact, formControl, updateModalDimensions, endLeftX, endRightX, endBottomY, landscapeMediaQuery, updateDimensions, spaceoutThumbs, updateEndBottomY, updateDimensionsNoMargins, getThumbWidthWithoutMargin, setupNavbar, updateEndTopY, getEndTopY, updateWIPDimensions, hideScrollBar, showScrollBar, enableStickyNavbar, setupDynamicLinks, setupNavbarEvents, autoScrollNow, showWip, hideWip } from "./navbar.js";
+import {SCROLL_DURATION, formControl, landscapeMediaQuery, setupNavbar, 
+    updateWIPDimensions, hideScrollBar, 
+  showScrollBar, enableStickyNavbar, setupDynamicLinks, setupNavbarEvents, 
+  autoScrollNow, showWip, hideWip } from "./navbar.js";
 
 
-// svg.style.visibility = "hidden"; // Hide SVG initially
+  
+
+
 const thumbnailsContainer = document.querySelector("#thumbnails");
-
+let screenWidthHalved = svg.viewBox.baseVal.width / 2;
+let screenHeightHalved = svg.viewBox.baseVal.height / 2;
+let endLeftX, endRightX, endBottomY, endTopY;
 // const svg = document.querySelector("#svg");
 
 
@@ -296,117 +303,129 @@ window.addEventListener("load", function () {
 // Thumbnails positioning based on window size
 
 
-// function getComputedStyleValue(element, property) {
+function getComputedStyleValue(element, property) {
 
-//   return parseInt(window.getComputedStyle(element).getPropertyValue(property));
-// }
+  return parseInt(window.getComputedStyle(element).getPropertyValue(property));
+}
 
-// function getThumbWidthWithMargin() {
-//   const computedStyle = window.getComputedStyle(domElements.thumbElements[0]);
-//   const thumbWidth = parseFloat(computedStyle.getPropertyValue("width"));
-//   const thumbMargin = parseFloat(
-//     computedStyle.getPropertyValue("margin-right")
-//   );
 
-//   return thumbWidth + thumbMargin * 2;
-// }
+
+function getThumbWidthWithMargin() {
+  const computedStyle = window.getComputedStyle(domElements.thumbElements[0]);
+  const thumbWidth = parseFloat(computedStyle.getPropertyValue("width"));
+  const thumbMargin = parseFloat(
+    computedStyle.getPropertyValue("margin-right")
+  );
+
+  return thumbWidth + thumbMargin * 2;
+}
+
+
 
 // const svg = document.querySelector("svg");
-// let thumbWidth = getComputedStyleValue(
-//   document.querySelector(".thumbShape"),
-//   "width"
-// );
-// let screenWidthHalved = svg.viewBox.baseVal.width / 2;
-// let screenHeightHalved = svg.viewBox.baseVal.height / 2;
-// let endLeftX, endRightX, endBottomY; //getEndTopY(),
+let thumbWidth = getComputedStyleValue(
+  document.querySelector(".thumbShape"),
+  "width"
+);
+ //getEndTopY(),
 
 // const landscapeMediaQuery = window.matchMedia(
 //   "(orientation: landscape) and (max-width: 991.98px) and (max-height: 600px)"
 // );
+
+
+const totalThumbWidth = getThumbWidthWithMargin();
+
 // const totalThumbWidth = getThumbWidthWithMargin();
 
-// // Functions to position thumbnails when media query is satisfied
+
+
+// function getEndTopY() {
+//   const updatedEndTopY = calculateEndTopY(); // Ensure `calculateEndTopY()` exists
+//   console.log("✅ getEndTopY value:", updatedEndTopY);
+//   return updatedEndTopY;
+// }
+
+// // // Functions to position thumbnails when media query is satisfied
 // function updateEndTopY() {
-//   if (
-//     window.matchMedia("(orientation: landscape) and (max-width: 991.98px)")
-//       .matches
-//   ) {
-//     getEndTopY = window.innerHeight * 1.275; // Adjust multiplier for this condition
-//   } else {
-//     getEndTopY = window.innerHeight * 1.325; // Default multiplier
+//     if (window.matchMedia("(orientation: landscape) and (max-width: 991.98px)").matches) {
+//       endTopY = window.innerHeight * 1.275;
+//     } else {
+//       endTopY = window.innerHeight * 1.325;
+//     }
+//     return endTopY;
 //   }
-//   return getEndTopY();
-//   console.log("Updated endTopY:", endTopY);
-// }
-// function updateEndBottomY() {
-//   if (
-//     window.matchMedia("(orientation: landscape) and (max-width: 991.98px)")
-//       .matches
-//   ) {
-//     endBottomY = window.innerHeight * 1.275 + totalThumbWidth; // Adjust multiplier for this condition
-//   } else {
-//     endBottomY = window.innerHeight * 1.325 + totalThumbWidth;
-//   }
-//   return endBottomY;
-// }
+
+
+function updateEndBottomY() {
+  if (
+    window.matchMedia("(orientation: landscape) and (max-width: 991.98px)")
+      .matches
+  ) {
+    endBottomY = window.innerHeight * 1.275 + totalThumbWidth; // Adjust multiplier for this condition
+  } else {
+    endBottomY = window.innerHeight * 1.325 + totalThumbWidth;
+  }
+  return endBottomY;
+}
 
 // Listen for changes in the media query
 landscapeMediaQuery.addEventListener("change", updateEndTopY, updateEndBottomY);
 
-// function updateDimensions() {
-//   thumbWidth = Math.min(300, window.innerWidth / 6);
-//   screenWidthHalved = window.innerWidth / 2;
-//   screenHeightHalved = window.innerHeight / 2;
+function updateDimensions() {
+  thumbWidth = Math.min(300, window.innerWidth / 6);
+  screenWidthHalved = window.innerWidth / 2;
+  screenHeightHalved = window.innerHeight / 2;
 
-//   endLeftX = screenWidthHalved - totalThumbWidth;
+  endLeftX = screenWidthHalved - totalThumbWidth;
 
-//   updateEndTopY(); // ✅ Update the value first
-//   const updatedEndTopY = getEndTopY(); // ✅ Retrieve the updated value
+  updateEndTopY(); // ✅ Update the value first
+  const updatedEndTopY = getEndTopY(); // ✅ Retrieve the updated value
 
-//   if (updatedEndTopY === undefined) {
-//     console.error("Error: getEndTopY() returned undefined!");
-//     return; // Stop execution if the value is not set
-//   }
+  if (updatedEndTopY === undefined) {
+    console.error("Error: getEndTopY() returned undefined!");
+    return; // Stop execution if the value is not set
+  }
 
-//   const endTopY = updatedEndTopY; // ✅ Use a local variable, do not redeclare globally
+  const endTopY = updatedEndTopY; // ✅ Use a local variable, do not redeclare globally
 
-//   endRightX = screenWidthHalved;
-//   updateEndBottomY();
-// }
+  endRightX = screenWidthHalved;
+  updateEndBottomY();
+}
 
-// function spaceoutThumbs() {
-//   gsap.to("#software", {
-//     x: endLeftX,
-//     y: getEndTopY(),
-//     duration: 1,
-//     ease: "power2.out",
-//   });
-//   gsap.to("#photography", {
-//     x: endRightX,
-//     y: getEndTopY(),
-//     duration: 1,
-//     ease: "power2.out",
-//   });
-//   gsap.to("#diy", {
-//     x: endRightX,
-//     y: endBottomY,
-//     duration: 1,
-//     ease: "power2.out",
-//   });
-//   gsap.to("#motion", {
-//     x: endLeftX,
-//     y: endBottomY,
-//     duration: 1,
-//     ease: "power2.out",
-//   });
-// }
+function spaceoutThumbs() {
+  gsap.to("#software", {
+    x: endLeftX,
+    y: getEndTopY(),
+    duration: 1,
+    ease: "power2.out",
+  });
+  gsap.to("#photography", {
+    x: endRightX,
+    y: getEndTopY(),
+    duration: 1,
+    ease: "power2.out",
+  });
+  gsap.to("#diy", {
+    x: endRightX,
+    y: endBottomY,
+    duration: 1,
+    ease: "power2.out",
+  });
+  gsap.to("#motion", {
+    x: endLeftX,
+    y: endBottomY,
+    duration: 1,
+    ease: "power2.out",
+  });
+}
 
-// function getThumbWidthWithoutMargin() {
-//   const computedStyle = window.getComputedStyle(domElements.thumbElements[0]);
-//   const thumbWidth = parseFloat(computedStyle.getPropertyValue("width"));
+function getThumbWidthWithoutMargin() {
+  const computedStyle = window.getComputedStyle(domElements.thumbElements[0]);
+  const thumbWidth = parseFloat(computedStyle.getPropertyValue("width"));
 
-//   return thumbWidth;
-// }
+  return thumbWidth;
+}
 // function getThumbMargin() {
 //   const computedStyle = window.getComputedStyle(domElements.thumbElements[0]);
 //   const thumbMargin = parseFloat(
@@ -415,97 +434,97 @@ landscapeMediaQuery.addEventListener("change", updateEndTopY, updateEndBottomY);
 //   return thumbMargin;
 // }
 
-// function updateDimensionsNoMargins() {
-//   setTimeout(() => {
-//     thumbWidth = Math.min(300, window.innerWidth / 6);
-//     console.log("thumbWidth: ", thumbWidth);
-//     screenWidthHalved = window.innerWidth / 2;
-//     screenHeightHalved = window.innerHeight / 2;
-//     const widthThumb = getThumbWidthWithoutMargin();
-//     const marginWidth = getThumbMargin();
+function updateDimensionsNoMargins() {
+  setTimeout(() => {
+    thumbWidth = Math.min(300, window.innerWidth / 6);
+    console.log("thumbWidth: ", thumbWidth);
+    screenWidthHalved = window.innerWidth / 2;
+    screenHeightHalved = window.innerHeight / 2;
+    const widthThumb = getThumbWidthWithoutMargin();
+    const marginWidth = getThumbMargin();
 
-//     updateEndTopY();
+    updateEndTopY();
 
-//     const updatedEndTopY = getEndTopY(); // ✅ Fetch dynamically
-//     if (updatedEndTopY === undefined) {
-//       console.error("Error: getEndTopY() returned undefined!");
-//       return; 
-//     }
+    const updatedEndTopY = getEndTopY(); // ✅ Fetch dynamically
+    if (updatedEndTopY === undefined) {
+      console.error("Error: getEndTopY() returned undefined!");
+      return; 
+    }
 
-//     const endTopY = updatedEndTopY + marginWidth - 15; 
-//     endLeftX = screenWidthHalved - (widthThumb + marginWidth);
-//     endRightX = screenWidthHalved - marginWidth;
-//     endBottomY = updateEndBottomY() - marginWidth - 15;
+    const endTopY = updatedEndTopY + marginWidth - 15; 
+    endLeftX = screenWidthHalved - (widthThumb + marginWidth);
+    endRightX = screenWidthHalved - marginWidth;
+    endBottomY = updateEndBottomY() - marginWidth - 15;
 
-//     collectThumbs();
-//     updateModalDimensions(endTopY);
-//     formControl(endTopY);
-//   }, 450); // delay to ensure scrollbar removal takes effect
-// }
+    collectThumbs();
+    updateModalDimensions(endTopY);
+    formControl(endTopY);
+  }, 450); // delay to ensure scrollbar removal takes effect
+}
 //when 'ME' is clicked this bunches the thumbnails together for background to modal
-// function collectThumbs() {
-//   gsap.to("#software", {
-//     scale: 1,
-//     x: endLeftX,
-//     y: getEndTopY(),
-//     duration: 1,
-//     ease: "power2.out",
-//   });
-//   gsap.to("#photography", {
-//     scale: 1,
-//     x: endRightX,
-//     y: getEndTopY(),
-//     duration: 1,
-//     ease: "power2.out",
-//   });
-//   gsap.to("#diy", {
-//     scale: 1,
-//     x: endRightX,
-//     y: endBottomY,
-//     duration: 1,
-//     ease: "power2.out",
-//   });
-//   gsap.to("#motion", {
-//     scale: 1,
-//     x: endLeftX,
-//     y: endBottomY,
-//     duration: 1,
-//     ease: "power2.out",
-//   });
-// }
+function collectThumbs() {
+  gsap.to("#software", {
+    scale: 1,
+    x: endLeftX,
+    y: getEndTopY(),
+    duration: 1,
+    ease: "power2.out",
+  });
+  gsap.to("#photography", {
+    scale: 1,
+    x: endRightX,
+    y: getEndTopY(),
+    duration: 1,
+    ease: "power2.out",
+  });
+  gsap.to("#diy", {
+    scale: 1,
+    x: endRightX,
+    y: endBottomY,
+    duration: 1,
+    ease: "power2.out",
+  });
+  gsap.to("#motion", {
+    scale: 1,
+    x: endLeftX,
+    y: endBottomY,
+    duration: 1,
+    ease: "power2.out",
+  });
+}
 
 
-// function updateModalDimensions(endTopY) {
-//   const modalBox = document.querySelector(".modalbox .box");
+function updateModalDimensions(endTopY) {
+  const modalBox = document.querySelector(".modalbox .box");
 
-//   if (!modalBox || !domElements.thumbElements[0]) return;
+  if (!modalBox || !domElements.thumbElements[0]) return;
 
-//   const thumbWidthWithoutMargin = getThumbWidthWithoutMargin();
+  const thumbWidthWithoutMargin = getThumbWidthWithoutMargin();
 
-//   // Calculate new width and height for the modal box
-//   const newWidth = Math.max(thumbWidthWithoutMargin * 2 + 4, 300);
-//   const newHeight = newWidth; // Assuming we want a square modal
+  // Calculate new width and height for the modal box
+  const newWidth = Math.max(thumbWidthWithoutMargin * 2 + 4, 300);
+  const newHeight = newWidth; // Assuming we want a square modal
 
-//   // Update modal dimensions
-//   modalBox.style.width = `${newWidth}px`;
-//   modalBox.style.height = `${newHeight}px`;
+  // Update modal dimensions
+  modalBox.style.width = `${newWidth}px`;
+  modalBox.style.height = `${newHeight}px`;
 
-//   // Calculate the center of the screen
-//   const centerX = window.innerWidth / 2 + 6; //refining positioning
+  // Calculate the center of the screen
+  const centerX = window.innerWidth / 2 + 6; //refining positioning
 
-//   // Calculate the new left position to center the modal box
-//   const newLeft = centerX - newWidth / 2;
+  // Calculate the new left position to center the modal box
+  const newLeft = centerX - newWidth / 2;
 
-//   // Use the passed endTopY for the new top position
-//   const newTop = getEndTopY() + 12.5; //  works for alignment on y axis
+  // Use the passed endTopY for the new top position
+  const newTop = getEndTopY() + 12.5; //  works for alignment on y axis
 
-//   // Update modal position
-//   modalBox.style.position = "absolute";
-//   modalBox.style.left = `${newLeft}px`;
-//   modalBox.style.top = `${newTop}px`;
+  // Update modal position
+  modalBox.style.position = "absolute";
+  modalBox.style.left = `${newLeft}px`;
+  modalBox.style.top = `${newTop}px`;
 
-//   // modalBox.style.display = "block";
-// }
+  // modalBox.style.display = "block";
+}
 
 function getScrollbarWidth() {
   // Create a temporary div element
@@ -794,11 +813,11 @@ function showForm() {
 }
 
 // Displays Welcome message
-// function showStatementContact() {
-//   updateDimensionsNoMargins();
-//   document.getElementById("statementContact").style.display = "block";
-//   document.getElementById("contactForm").style.display = "none";
-// }
+function showStatementContact() {
+  updateDimensionsNoMargins();
+  document.getElementById("statementContact").style.display = "block";
+  document.getElementById("contactForm").style.display = "none";
+}
 
 // // Displays WIP message
 // function showWip() {
